+++ doc/CHANGELOG.md
@@ -19,6 +19,35 @@ Further documentation can be found in:
    It is also available at http://plse.cs.washington.edu/daikon/download/api/ .
 
 
+Version 5.9.0 (??)
+=============================
+
+Daikon supports the following new invariants, over strings and sequences of
+strings; all of them are enabled by default:
+ * IsUrl
+ * FixedLengthString
+ * IsNumeric
+ * IsEmail
+ * IsDateYYYYMMDD
+ * IsDateDDMMYYYY
+ * IsDateMMDDYYYY
+ * IsTimeOfDay
+ * IsTimeOfDayWithSeconds
+ * IsTimeOfDayAMPM
+ * IsTimestampYYYYMMHHThhmmssmm
+ * SequenceFixedLengthString
+ * SequenceStringElementsAreUrl
+ * SequenceStringElementsAreNumeric
+ * SequenceStringElementsAreEmail
+ * SequenceStringElementsAreDateYYYYMMDD
+ * SequenceStringElementsAreDateDDMMYYYY
+ * SequenceStringElementsAreDateMMDDYYYY
+ * SequenceStringElementsAreTimeOfDay
+ * SequenceStringElementsAreTimeOfDayWithSeconds
+ * SequenceStringElementsAreTimeOfDayAMPM
+ * SequenceStringElementsAreTimestampYYYYMMHHThhmmssmm
+
+
 Version 5.8.20 (May 14, 2024)
 =============================